+++ src/formatting/mod.rs
@@ -1,4 +1,4 @@
-mod formatter;
+mod renderer;
 
 // Re-export all public symbols
-pub use formatter::*;
+pub use renderer::*;
